requirements.txt
asgiref==3.8.1
Django==5.1.2
django-background-tasks==1.2.8
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
